Implement rest controllers and basic React components

+++ EShopRedesign/backend/src/main/java/mk/ukim/finki/nvd/backend/repository/ShoppingCartRepository.java
@@ -3,5 +3,6 @@ package mk.ukim.finki.nvd.backend.repository;
 import mk.ukim.finki.nvd.backend.model.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
+
 }
